chatbot-project/requirements.txt
openai==1.53.0
SpeechRecognition==3.11.0
pyserial==3.5
numpy==1.25.0
opencv-python==4.10.0.84
pyrealsense2==2.55.1.6486
pyttsx3>=2.90
rich>=10.12.0
ultralytics==8.3.27

python-dotenv>=1.0.1
